% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aus_athletes}
\alias{Aus_athletes}
\title{Australian athletes}
\format{
An object of class \code{data.frame} with 202 rows and 13 columns.
}
\usage{
Aus_athletes
}
\description{
This dataset is equivalent to \code{ais} from the \code{DAAG} package.
}
\examples{
# none yet
}
\references{
Telford, R.D. and Cunningham, R.B. 1991. Sex, sport and body-size dependency of hematology in
highly trained athletes. Medicine and Science in Sports and Exercise 23: 788-794.
}
\keyword{datasets}
