% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Catalan_elections}
\alias{Catalan_elections}
\title{Results from Catalan regional elections (1980-2015)}
\format{
A tibble with 20764 rows and 4 variables:
\describe{
\item{\code{Municipality}}{}
\item{\code{Year}}{}
\item{\code{Option}}{The voter option; either "Indy" or "Unionist"}
\item{\code{Percent}}{The percentage of the voters choosing the given option}
}
}
\usage{
Catalan_elections
}
\description{
Data from Catalan regional elections for 949 municipalities, from 11 elections spanning the years
1980-2015. The data was obtained and processed from Idescat.cat by Marc Belzunces (Twitter: @marcbeldata).
}
\keyword{datasets}
