% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggridges.R
\docType{package}
\name{ggridges-package}
\alias{ggridges}
\alias{ggridges-package}
\title{Ridgeline plots with ggplot2}
\description{
Please see the package vignettes for usage instructions. For a quick start,
check out the examples for \code{\link[=geom_density_ridges]{geom_density_ridges()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wilkelab.org/ggridges/}
  \item Report bugs at \url{https://github.com/wilkelab/ggridges/issues}
}

}
\author{
\strong{Maintainer}: Claus O. Wilke \email{wilke@austin.utexas.edu} (\href{https://orcid.org/0000-0002-7470-9261}{ORCID})

}
\keyword{internal}
