% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggridges.R
\docType{package}
\name{ggridges}
\alias{ggridges}
\alias{ggridges-package}
\title{Ridgeline plots with ggplot2}
\description{
Please see the package vignettes for usage instructions. For a quick start,
check out the examples for \code{\link[=geom_density_ridges]{geom_density_ridges()}}.
}
