% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lincoln_weather}
\alias{lincoln_weather}
\title{Weather in Lincoln, Nebraska in 2016.}
\format{A tibble with 366 rows and 24 variables:
\describe{
\item{CST}{Day of the measurement}
\item{Max Temperature \link{F}}{Maximum temperature in Fahrenheit}
\item{Mean Temperature \link{F}}{Maximum temperature in Fahrenheit}
\item{Month}{The month in which the measurement was taken}
}}
\usage{
lincoln_weather
}
\description{
A dataset containing weather information from Lincoln, Nebraska, from 2016.
Originally downloaded from Weather Underground by Austin Wehrwein, http://austinwehrwein.com/.
The variables are as follows:
}
\keyword{datasets}
