% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\docType{data}
\name{position_points_sina}
\alias{position_points_sina}
\alias{PositionPointsSina}
\title{Randomly distribute points in a ridgeline plot between baseline and ridgeline}
\usage{
position_points_sina(rel_min = 0.02, rel_max = 0.98, seed = NULL)
}
\arguments{
\item{rel_min}{The relative minimum value at which a point can be placed.}

\item{rel_max}{The relative maximum value at which a point can be placed.}

\item{seed}{See \code{\link{position_points_jitter}}.}
}
\description{
Randomly distribute points in a ridgeline plot between baseline and ridgeline
}
\examples{
ggplot(iris, aes(x = Sepal.Length, y = Species)) +
  geom_density_ridges(jittered_points = TRUE, position = "points_sina", alpha = 0.7)
}
\seealso{
\code{\link{position_points_jitter}}, \code{\link{position_raincloud}}
}
\keyword{datasets}
