% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\docType{data}
\name{position_raincloud}
\alias{position_raincloud}
\alias{PositionRaincloud}
\title{Randomly jitter the points below a ridgeline plot}
\usage{
position_raincloud(width = 0, height = 0.4, ygap = 0.05, seed = NULL)
}
\arguments{
\item{width}{Width for horizonal jittering. By default set to 0.}

\item{height}{Total height of point cloud. By default 0.4.}

\item{ygap}{Vertical gap between ridgeline baseline and point cloud.}

\item{seed}{Random seed. See \code{\link{position_points_jitter}}.}
}
\description{
Randomly jitter the points below a ridgeline plot
}
\examples{
ggplot(iris, aes(x = Sepal.Length, y = Species)) +
  geom_density_ridges(jittered_points = TRUE, position = "raincloud", alpha = 0.7)
}
\seealso{
position_points_jitter

position_points_sina
}
\keyword{datasets}
