% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-cyclical.R
\docType{data}
\name{scale_cyclical}
\alias{scale_colour_cyclical}
\alias{scale_color_cyclical}
\alias{scale_fill_cyclical}
\alias{cyclical_scale}
\alias{ScaleCyclical}
\title{Create a discrete scale that cycles between values}
\usage{
scale_colour_cyclical(..., values)

scale_fill_cyclical(..., values)

cyclical_scale(aesthetics, values, name = waiver(), breaks = waiver(),
  labels = waiver(), limits = NULL, expand = waiver(),
  na.translate = TRUE, na.value = NA, drop = TRUE, guide = "legend",
  position = "left")
}
\description{
Create a discrete scale that cycles between values

Cyclical scale constructor
}
\examples{
ggplot(mtcars, aes(mpg, wt, color = factor(carb))) +
  geom_point() +
  scale_color_cyclical(values = c("blue", "green"))


}
\keyword{datasets}
