% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-scales.R
\name{scale_discrete_manual}
\alias{scale_discrete_manual}
\title{Generic discrete manual scale}
\usage{
scale_discrete_manual(aesthetics, values, ...)
}
\arguments{
\item{aesthetics}{The aesthetics for which this scale should be used}

\item{values}{List of values to be used as palette}

\item{...}{Other parameters handed off to \link{discrete_scale}}
}
\description{
Generic discrete manual scale.
}
