% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_binline}
\alias{stat_binline}
\title{Stat for histogram joyplots}
\usage{
stat_binline(mapping = NULL, data = NULL, geom = "joy",
  position = "identity", ..., binwidth = NULL, bins = NULL,
  center = NULL, boundary = NULL, breaks = NULL, closed = c("right",
  "left"), pad = TRUE, draw_baseline = TRUE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{draw_baseline}{If \code{FALSE}, removes lines along 0 counts. Default is \code{TRUE}.}
}
\description{
Works like \code{stat_bin} except that the output is a ridgeline describing the histogram rather than
a set of counts.
}
\examples{
ggplot(iris, aes(x = Sepal.Length, y = Species, group = Species, fill = Species)) +
  geom_joy(stat = "binline", bins = 20, scale = 2.2) +
  scale_y_discrete(expand = c(0.01, 0)) +
  scale_x_continuous(expand = c(0.01, 0)) +
  theme_joy()

ggplot(iris, aes(x = Sepal.Length, y = Species, group = Species, fill = Species)) +
  stat_binline(bins = 20, scale = 2.2, draw_baseline = FALSE) +
  scale_y_discrete(expand = c(0.01, 0)) +
  scale_x_continuous(expand = c(0.01, 0)) +
  theme_joy()

require(ggplot2movies)
ggplot(movies[movies$year>1989,], aes(x = length, y = year, group = year)) +
  stat_binline(scale = .9, size = 0.25) +
    theme_joy() +
    scale_x_continuous(limits = c(1, 200), expand = c(0.01, 0)) +
    scale_y_reverse(expand = c(0.01, 0))

}
