% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_joy}
\alias{stat_joy}
\title{Stat for density joyplots}
\usage{
stat_joy(mapping = NULL, data = NULL, geom = "joy",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, bandwidth = NULL, ...)
}
\description{
This stat is the default stat used by \code{geom_joy}. It is very similar to \code{stat_density},
however there are a few differences. Most importantly, the density bandwidth is chosen across
the entire dataset.
}
