% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_joy}
\alias{theme_joy}
\title{A custom theme specifically for use with joyplots}
\usage{
theme_joy(font_size = 14, font_family = "", line_size = 0.5)
}
\arguments{
\item{font_size}{Overall font size. Default is 14.}

\item{font_family}{Default font family.}

\item{line_size}{Default line size.}
}
\value{
The theme.
}
\description{
This theme has some special modifications that make joyplots look better, such as properly aligned y axis labels.
}
\examples{
# still to do
}
