% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{geom_joy}
\alias{geom_joy}
\alias{GeomJoy}
\title{Joy plot based on ridgelines}
\usage{
geom_joy(mapping = NULL, data = NULL, stat = "density",
  position = "identity", na.rm = FALSE, show.legend = NA, scale = 1.8,
  inherit.aes = TRUE, ...)
}
\description{
`geom_joy` arranges multiple density plots in a staggered fashion, as in the cover of the famous Joy Division album.
}
\examples{
ggplot(iris, aes(x=Sepal.Length, y=Species, group=Species, height = ..density..)) +
  geom_joy() +
  scale_y_discrete(expand=c(0.01, 0)) +
  scale_x_continuous(expand=c(0.01, 0)) +
  theme_joy()


# set the scale argument in `geom_joy2()` to determine how much overlap there is among the plots
ggplot(diamonds, aes(x=price, y=cut, group=cut, height=..density..)) +
  geom_joy(scale=4) +
  scale_y_discrete(expand=c(0.01, 0)) +
  scale_x_continuous(expand=c(0.01, 0)) +
  theme_joy()

# the same figure with fun colors
ggplot(diamonds, aes(x=price, y=cut, fill=cut, height=..density..)) +
  geom_joy(scale=4) +
  scale_y_discrete(expand=c(0.01, 0)) +
  scale_x_continuous(expand=c(0.01, 0)) +
  scale_fill_brewer(palette = 4) +
  theme_joy() + theme(legend.position="none")
}
\keyword{datasets}
