% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{geom_joy2}
\alias{geom_joy2}
\alias{GeomJoy2}
\title{Joy plot based on closed polygons}
\usage{
geom_joy2(mapping = NULL, data = NULL, stat = "density",
  position = "identity", na.rm = FALSE, show.legend = NA, scale = 1.8,
  rel_min_height = 0, inherit.aes = TRUE, ...)
}
\description{
`geom_joy2` is similar to `geom_joy` but draws closed polygons rather than ridgelines.
}
\examples{
ggplot(iris, aes(x=Sepal.Length, y=Species, group=Species, height = ..density..)) +
  geom_joy2() +
  scale_y_discrete(expand=c(0.01, 0)) +
  scale_x_continuous(expand=c(0.01, 0)) +
  theme_joy()


# set the scale argument in `geom_joy2()` to determine how much overlap there is among the plots
ggplot(diamonds, aes(x=price, y=cut, group=cut, height=..density..)) +
  geom_joy2(scale=4) +
  scale_y_discrete(expand=c(0.01, 0)) +
  scale_x_continuous(expand=c(0.01, 0)) +
  theme_joy()
}
\keyword{datasets}
