% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_ridgeline}
\alias{geom_ridgeline}
\title{Plot a ridgeline (line with filled area underneath)}
\usage{
geom_ridgeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Plot a ridgeline (line with filled area underneath)
}
\examples{

d <- data.frame(x = rep(1:5, 3), y = c(rep(0, 5), rep(1, 5), rep(3, 5)),
                height = c(0, 1, 3, 4, 0, 1, 2, 3, 5, 4, 0, 5, 4, 4, 1))
ggplot(d, aes(x, y, height = height, group = y))) + geom_ridgeline()

}
