% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-scales.R
\name{scale_point}
\alias{scale_point_shape}
\alias{scale_point_color_hue}
\alias{scale_point_fill_hue}
\alias{scale_point_shape_discrete}
\alias{scale_point_color_discrete}
\alias{scale_point_fill_discrete}
\title{Scales for point aesthetics}
\usage{
scale_point_shape(..., solid = TRUE)

scale_point_color_hue(..., h = c(0, 360) + 15, c = 100, l = 65,
  h.start = 0, direction = 1, na.value = "grey50",
  aesthetic = "point_color")

scale_point_fill_hue(...)

scale_point_shape_discrete(..., solid = TRUE)

scale_point_color_discrete(..., h = c(0, 360) + 15, c = 100, l = 65,
  h.start = 0, direction = 1, na.value = "grey50",
  aesthetic = "point_color")

scale_point_fill_discrete(...)
}
\description{
Scales for point aesthetics
}
